+++ src/stories/auto-refresh.js
@@ -1,6 +1,6 @@
 // sample auto refresh component to screen grab gifs
-import React from "react"
-import ReactSpeedometer from "../../src/"
+import React from 'react'
+import ReactSpeedometer from '../../src/'
 
 class AutoRefresh extends React.Component {
   constructor(props) {
@@ -11,7 +11,7 @@ class AutoRefresh extends React.Component {
   }
   componentDidMount = () => {
     setInterval(() => {
-      this.setState((prevState) => ({
+      this.setState(prevState => ({
         opacity: prevState.opacity === 1 ? 0.9 : 1,
       }))
     }, 3333)
@@ -21,21 +21,21 @@ class AutoRefresh extends React.Component {
     return (
       <div
         style={{
-          background: "#2e3440",
-          background: "#2a2744",
+          background: '#2e3440',
+          background: '#2a2744',
         }}
       >
         <div
           style={{
-            display: "flex",
-            justifyContent: "center",
+            display: 'flex',
+            justifyContent: 'center',
             // alignItems: "center",
-            height: "231px",
+            height: '231px',
           }}
         >
           <div
             style={{
-              margin: "7px 17px",
+              margin: '7px 17px',
             }}
           >
             <ReactSpeedometer
@@ -43,44 +43,44 @@ class AutoRefresh extends React.Component {
               segments={3}
               value={471}
               forceRender={true}
-              textColor={"#d8dee9"}
-              segmentColors={["#FF9933", "#FFF", "#138808"]}
+              textColor={'#d8dee9'}
+              segmentColors={['#FF9933', '#FFF', '#138808']}
               needleColor={`rgba(94, 129, 172, ${this.state.opacity})`}
               needleTransition="easeElastic"
               needleTransitionDuration={3333}
-              currentValueText={"Current Value: ${value}"}
+              currentValueText={'Current Value: ${value}'}
             />
           </div>
           <div
             style={{
-              margin: "7px 17px",
+              margin: '7px 17px',
             }}
           >
             <ReactSpeedometer
               value={333}
               segments={77}
               maxSegmentLabels={1}
-              textColor={"#d8dee9"}
+              textColor={'#d8dee9'}
               needleColor={`rgba(94, 129, 172, ${this.state.opacity})`}
               forceRender={true}
             />
           </div>
           <div
             style={{
-              margin: "7px 17px",
+              margin: '7px 17px',
             }}
           >
             <ReactSpeedometer
               segments={5}
               segmentColors={[
-                "#bf616a",
-                "#d08770",
-                "#ebcb8b",
-                "#a3be8c",
-                "#b48ead",
+                '#bf616a',
+                '#d08770',
+                '#ebcb8b',
+                '#a3be8c',
+                '#b48ead',
               ]}
               maxSegmentLabels={5}
-              textColor={"#d8dee9"}
+              textColor={'#d8dee9'}
               needleColor="#eceff4"
               needleTransition="easeElastic"
               needleTransitionDuration={3333}
@@ -91,8 +91,8 @@ class AutoRefresh extends React.Component {
         </div>
         <div
           style={{
-            display: "flex",
-            justifyContent: "center",
+            display: 'flex',
+            justifyContent: 'center',
             // alignItems: "center",
           }}
         >
@@ -102,40 +102,40 @@ class AutoRefresh extends React.Component {
               needleHeightRatio={0.7}
               value={this.state.opacity === 1 ? 777 : 931}
               currentValueText="Happiness Level"
-              startColor={"#ff927e"}
+              startColor={'#ff927e'}
               customSegmentLabels={[
                 {
-                  text: "Very Bad",
-                  position: "INSIDE",
-                  color: "#555",
+                  text: 'Very Bad',
+                  position: 'INSIDE',
+                  color: '#555',
                 },
                 {
-                  text: "Bad",
-                  position: "INSIDE",
-                  color: "#555",
+                  text: 'Bad',
+                  position: 'INSIDE',
+                  color: '#555',
                 },
                 {
-                  text: "Ok",
-                  position: "INSIDE",
-                  color: "#555",
-                  fontSize: "19px",
+                  text: 'Ok',
+                  position: 'INSIDE',
+                  color: '#555',
+                  fontSize: '19px',
                 },
                 {
-                  text: "Good",
-                  position: "INSIDE",
-                  color: "#555",
+                  text: 'Good',
+                  position: 'INSIDE',
+                  color: '#555',
                 },
                 {
-                  text: "Very Good",
-                  position: "INSIDE",
-                  color: "#555",
+                  text: 'Very Good',
+                  position: 'INSIDE',
+                  color: '#555',
                 },
               ]}
               ringWidth={47}
               needleTransitionDuration={3333}
               needleTransition="easeElastic"
-              needleColor={"#90f2ff"}
-              textColor={"#d8dee9"}
+              needleColor={'#90f2ff'}
+              textColor={'#d8dee9'}
             />
           </div>
           <div>
@@ -144,30 +144,30 @@ class AutoRefresh extends React.Component {
               needleHeightRatio={0.7}
               value={this.state.opacity === 1 ? 747 : 871}
               customSegmentStops={[0, 250, 750, 1000]}
-              segmentColors={["#9399ff", "#14ffec", "#00bbf0"]}
+              segmentColors={['#9399ff', '#14ffec', '#00bbf0']}
               currentValueText="How are you?"
               customSegmentLabels={[
                 {
-                  text: "Good",
-                  position: "OUTSIDE",
-                  color: "#d8dee9",
+                  text: 'Good',
+                  position: 'OUTSIDE',
+                  color: '#d8dee9',
                 },
                 {
-                  text: "Great",
-                  position: "OUTSIDE",
-                  color: "#d8dee9",
+                  text: 'Great',
+                  position: 'OUTSIDE',
+                  color: '#d8dee9',
                 },
                 {
-                  text: "Awesome!",
-                  position: "OUTSIDE",
-                  color: "#d8dee9",
+                  text: 'Awesome!',
+                  position: 'OUTSIDE',
+                  color: '#d8dee9',
                 },
               ]}
               ringWidth={47}
               needleTransitionDuration={3333}
               needleTransition="easeElastic"
-              needleColor={"#a7ff83"}
-              textColor={"#d8dee9"}
+              needleColor={'#a7ff83'}
+              textColor={'#d8dee9'}
             />
           </div>
         </div>