+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.14.0
+- `valueTextFontWeight` config/prop to control font weight of current value
+- *dev changes*: prettier changes: `singleQuote: true`, `arrowParens: avoid`. `prettier` upgraded to `v2`
+- *dev changes*: `jest v26`
+
 ## 0.13.1
 - ignore `coverage` folder from npm bundle